Added reset feature on zone departure
