+++ Assets/Prefabs/RespawnZone.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0c2de9c141cbd0340b391adeac427025
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
