+++ Assets/Prefabs/Small Box.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 1dc0a9a155d4d074490f17094bc5b805
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
