+++ Assets/Scripts/Reveal.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f1bf15d136b4ba94f9979e46a405a5b1
+guid: dec585bdb98acdf46921180631dd5181
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
