+++ Assets/Scripts/StateTracker.cs
@@ -1,50 +1,43 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.InputSystem;
 
 public class StateTracker : MonoBehaviour
 {
-    public float visibilityRate = 0.01f;
-    float childOpacity = 0;
-    bool reveal;
-
+    public GameObject prefab;
+    bool resetLive;
     // Start is called before the first frame update
     void Start()
     {
-        SetChildrenAlpha(0);
+
     }
 
     // Update is called once per frame
     void Update()
     {
-        if (reveal && childOpacity < 1) {
-            childOpacity += visibilityRate;
-            SetChildrenAlpha(childOpacity);
-        } else if (!reveal && childOpacity > 0) {
-            childOpacity -= visibilityRate;
-            SetChildrenAlpha(childOpacity);
-        }
+
     }
 
-    void OnTriggerEnter2D(Collider2D collision) {
+    private void OnTriggerEnter2D(Collider2D collision) {
         if (collision.gameObject.CompareTag("Player")) {
-            reveal = true;
+            resetLive = true;
         }
     }
 
-    void OnTriggerExit2D(Collider2D collision) {
+    private void OnTriggerExit2D(Collider2D collision) {
         if (collision.gameObject.CompareTag("Player")) {
-            reveal = false;
+            ResetState();
+            resetLive = false;
         }
     }
 
-    void SetChildrenAlpha(float alpha) {
+    void ResetState() {
         foreach (Transform child in gameObject.GetComponentsInChildren<Transform>()) {
             if (child != this.transform) {
-                var mat = child.GetComponent<Renderer>().material;
-                Color newColor = new(mat.color.r, mat.color.g, mat.color.b, alpha);
-                mat.SetColor("_Color", newColor);
+                Destroy(child.gameObject);
             }
         }
+        Instantiate(prefab, gameObject.transform);
     }
 }
